fix(build): make the sources compile under the debug preset
With the preset fixed, `cmake --preset debug` configures but its flag set
(-pedantic-errors -Werror -Wconversion -Wsign-conversion -Wshadow ...)
had never actually been run against this tree, and it did not compile.
17 errors across seven files, none of them previously visible.

  -Wsign-conversion (7): int counts passed where size_t is expected —
    qsort element counts and calloc member counts. Cast explicitly. In
    subtitle_convert.c the factors are cast individually,
    calloc((size_t)w * (size_t)h, 1), which also removes a latent int
    overflow before the widening.
  -Wimplicit-int-float-conversion (2): int64_t operands feeding double
    arithmetic in the crop sampler and the fps readout. Cast explicitly.
  -Wimplicit-int-conversion (1): the int result of a byte-pair shift
    narrowed to the uint16_t return type.
  -Wimplicit-float-conversion (1): a double literal passed to leptonica's
    l_float32 parameter, now 0.1F (uppercase, per
    readability-uppercase-literal-suffix).
  -Wmissing-format-attribute (3): format(printf, ...) attributes on
    vmav_cmd_argf and the two SVT log callbacks, matching how ui.h
    already annotates its printf-style functions. These are real: without
    them the compiler cannot check format strings at call sites.

Every change makes an existing implicit conversion explicit, so none of
them alters behaviour. Confirmed: `cmake --preset debug` now builds
clean, vmav_tests passes under it, and the 60s reference encode is
bit-identical to before these edits — same video and audio stream MD5s,
same byte count, same keyframe layout.

+++ src/media_crop/media_crop.c
@@ -190,7 +190,7 @@ CropInfo get_crop_info(const char *path) {
   int sample_count = 0;
 
   for (int s = 0; s < CROP_SAMPLES && sample_count < CROP_SAMPLES; s++) {
-    int64_t target = duration / 10 + (int64_t)(duration * 0.8 / CROP_SAMPLES) * s;
+    int64_t target = duration / 10 + (int64_t)((double)duration * 0.8 / CROP_SAMPLES) * s;
 
     av_seek_frame(fmt_ctx, -1, target, AVSEEK_FLAG_BACKWARD);
     avcodec_flush_buffers(dec_ctx);
@@ -247,10 +247,10 @@ CropInfo get_crop_info(const char *path) {
   }
 
   if (sample_count > 0) {
-    qsort(tops, sample_count, sizeof(int), cmp_int);
-    qsort(bottoms, sample_count, sizeof(int), cmp_int);
-    qsort(lefts, sample_count, sizeof(int), cmp_int);
-    qsort(rights, sample_count, sizeof(int), cmp_int);
+    qsort(tops, (size_t)sample_count, sizeof(int), cmp_int);
+    qsort(bottoms, (size_t)sample_count, sizeof(int), cmp_int);
+    qsort(lefts, (size_t)sample_count, sizeof(int), cmp_int);
+    qsort(rights, (size_t)sample_count, sizeof(int), cmp_int);
 
     int mid = sample_count / 2;
     info.top = tops[mid];
